third_party/blink/web_tests/fast/frames/frameset-frameborder-overrides-border-expected.txt

Bug 17767: Frameset with frameborder=0 and border=N should not show a border

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS frames[0].offsetWidth is 150
PASS frames[1].offsetWidth is 150
PASS successfullyParsed is true

TEST COMPLETE
